pyproject.toml
[project]
name = "flash-atten-whl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.13"
dependencies = [
    "packaging>=25.0",
    "pygithub>=2.6.1",
    "typer>=0.15.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "marimo>=0.13.0",
]
